defect_detection/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
sqlalchemy==2.0.30
asyncpg==0.29.0
alembic==1.13.1
pydantic==2.7.1
pydantic-settings==2.2.1
python-dotenv==1.0.1
python-multipart==0.0.9

# ML
torch==2.3.0
torchvision==0.18.0
numpy==1.26.4
Pillow==10.3.0
scikit-learn==1.4.2
matplotlib==3.9.0
seaborn==0.13.2

# Testing
pytest==8.2.0
pytest-asyncio==0.23.6
httpx==0.27.0
